Moved byte-span hashing to hash utils
