+++ cpp/common/hash_utils.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include <array>
+#include <cstdint>
 #include <functional>
+#include <span>
 
 namespace tosca {
 
@@ -17,5 +19,26 @@ constexpr void HashRange(size_t& seed, It first, It last) {
     CombineHash(seed, *first);
   }
 }
+
+template <size_t N>
+struct HashBytes {
+  constexpr size_t operator()(std::span<const uint8_t, N> key) const noexcept {
+    // Hash initial bytes as multi-byte values.
+    using Word = uint32_t;  // HashCombine is tuned for 32-bit words and does not work well with 64-bit values.
+    constexpr size_t kWordSize = sizeof(Word);
+    constexpr size_t kNumFullWords = N / kWordSize;
+    size_t seed = 0;
+    if constexpr (kNumFullWords > 0) {
+      const Word* begin = reinterpret_cast<const Word*>(key.data());
+      HashRange(seed, begin, begin + kNumFullWords);
+    }
+
+    // Hash the rest as bytes.
+    if constexpr (N % kWordSize != 0) {
+      HashRange(seed, key.begin() + kNumFullWords * kWordSize, key.end());
+    }
+    return seed;
+  }
+};
 
 }  // namespace tosca
